+++ Src/Private/Get-AbrVbrServiceProvider.ps1
@@ -6,7 +6,7 @@ function Get-AbrVbrServiceProvider {
     .DESCRIPTION
         Documents the configuration of Veeam VBR in Word/HTML/Text formats using PScribo.
     .NOTES
-        Version:        0.5.1
+        Version:        0.5.2
         Author:         Jonathan Colon
         Twitter:        @jcolonfzenpr
         Github:         rebelinux
@@ -28,21 +28,34 @@ function Get-AbrVbrServiceProvider {
         try {
             if ((Get-VBRInstalledLicense | Where-Object {$_.Edition -in @("EnterprisePlus")}) -and (Get-VBRCloudProvider).count -gt 0) {
                 Section -Style Heading3 'Cloud Service Providers' {
-                    Paragraph "The following section provides a summary about configured Veeam Cloud Service Providers."
+                    Paragraph "The following section provides a summary about configured Veeam Cloud Connect Service Providers."
                     BlankLine
                     try {
                         $CloudProviders = Get-VBRCloudProvider
+                        $OutObj = @()
                         foreach ($CloudProvider in $CloudProviders) {
                             try {
-                                $OutObj = @()
-                                Write-PscriboMessage "Discovered $($CloudProvider.DNSName) Service Provider."
+                                Write-PscriboMessage "Discovered $($CloudProvider.DNSName) Service Provider summary information."
+                                if ($CloudProvider.ResourcesEnabled) {
+                                    $WanAcceleration = $CloudProvider.Resources.WanAccelerationEnabled
+                                }
+                                elseif ($CloudProvider.ReplicationResourcesEnabled) {
+                                    $WanAcceleration = $CloudProvider.ReplicationResources.WanAcceleratorEnabled
+                                }
+
+                                if ($CloudProvider.ResourcesEnabled) {
+                                    $VCCType = 'Cloud Repositories'
+                                }
+                                elseif ($CloudProvider.ReplicationResourcesEnabled) {
+                                    $VCCType = 'Replication Resources'
+                                }
+
                                 $inObj = [ordered] @{
                                     'DNS Name' = $CloudProvider.DNSName
-                                    'Repository Name' = $CloudProvider.Resources.RepositoryName
-                                    'Allocated Space' = "$([Math]::Round($CloudProvider.Resources.RepositoryAllocatedSpace / 1024))GB"
-                                    'Wan Acceleration' = $CloudProvider.Resources.WanAccelerationEnabled
+                                    'Cloud Connect Type' = $VCCType
+                                    'Wan Acceleration' = ConvertTo-TextYN $WanAcceleration
                                 }
-                                $OutObj = [pscustomobject]$inobj
+                                $OutObj += [pscustomobject]$inobj
                             }
                             catch {
                                 Write-PscriboMessage -IsWarning $_.Exception.Message
@@ -52,30 +65,48 @@ function Get-AbrVbrServiceProvider {
                         $TableParams = @{
                             Name = "Cloud Service Providers - $VeeamBackupServer"
                             List = $false
-                            ColumnWidths = 30, 30, 20, 20
+                            ColumnWidths = 35, 35, 30
                         }
 
                         if ($Report.ShowTableCaptions) {
                             $TableParams['Caption'] = "- $($TableParams.Name)"
                         }
-                        $OutObj | Table @TableParams
+                        $OutObj | Sort-Object -Property 'DNS Name' | Table @TableParams
                         if ($InfoLevel.Infrastructure.ServiceProvider -ge 2) {
                             try {
                                 foreach ($CloudProvider in $CloudProviders) {
                                     Section -Style Heading3 $CloudProvider.DNSName {
                                         $OutObj = @()
-                                        Write-PscriboMessage "Discovered $($CloudProvider.DNSName) Service Provider."
+                                        Write-PscriboMessage "Discovered $($CloudProvider.DNSName) Service Provider configuration information."
                                         $inObj = [ordered] @{
                                             'DNS Name' = $CloudProvider.DNSName
                                             'Ip Address' = $CloudProvider.IpAddress
                                             'Port' = $CloudProvider.Port
                                             'Credentials' = $CloudProvider.Credentials
                                             'Certificate Expiration Date' = $CloudProvider.Certificate.NotAfter
-                                            'BaaS Resources Enabled' = ConvertTo-TextYN $CloudProvider.ResourcesEnabled
-                                            'DRaaS Replication Enabled' = ConvertTo-TextYN $CloudProvider.ReplicationResourcesEnabled
                                             'Managed By Service Provider' = ConvertTo-TextYN $CloudProvider.IsManagedByProvider
-                                            'Description' = $CloudProvider.Description
                                         }
+                                        if ($CloudProvider.ResourcesEnabled) {
+                                            $inObj.add('BaaS Resources Enabled', (ConvertTo-TextYN $CloudProvider.ResourcesEnabled))
+                                            $inObj.add('Repository Name', $CloudProvider.Resources.RepositoryName)
+                                            $inObj.add('Allocated Resources', "$([Math]::Round($CloudProvider.Resources.RepositoryAllocatedSpace / 1024))GB")
+                                            $inObj.add('Description', $CloudProvider.Description)
+                                        }
+
+                                        if ($CloudProvider.ReplicationResourcesEnabled) {
+                                            $inObj.add('DRaaS Replication Enabled', (ConvertTo-TextYN $CloudProvider.ReplicationResourcesEnabled))
+                                            $inObj.add('Hardware Plan Name', $CloudProvider.ReplicationResources.HardwarePlanName)
+                                            $inObj.add('Allocated CPU Resources', $CloudProvider.ReplicationResources.CPU)
+                                            $inObj.add('Allocated Memory Resources', $CloudProvider.ReplicationResources.Memory)
+                                            $inObj.add('Allocated Datastore Resources', "$([Math]::Round($CloudProvider.ReplicationResources.Datastore.DatastoreAllocatedSpace / 1024))GB")
+                                            $inObj.add('Network Count', $CloudProvider.ReplicationResources.NetworkCount)
+                                            $inObj.add('Public Ip Enabled', (ConvertTo-TextYN $CloudProvider.ReplicationResources.PublicIpEnabled))
+                                            # if ($CloudProvider.ReplicationResources.PublicIpEnabled) {
+                                            #     $inObj.add('Public Ip', $CloudProvider.ReplicationResources.PublicIp)
+                                            # }
+                                            $inObj.add('Description', $CloudProvider.Description)
+                                        }
+
                                         $OutObj = [pscustomobject]$inobj
 
                                         $TableParams = @{